Sync client & server interfaces

+++ src/main/java/be/ephys/utilitas/feature/universal_interface/interface_adapters/InterfaceTileEntity.java
@@ -4,10 +4,12 @@ import be.ephys.utilitas.api.registry.UniversalInterfaceAdapter;
 import be.ephys.utilitas.base.helpers.WorldHelper;
 import be.ephys.utilitas.feature.link_wand.ItemLinker;
 import be.ephys.utilitas.feature.universal_interface.TileEntityInterface;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.ITextComponent;
@@ -17,7 +19,7 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 
 import javax.annotation.Nullable;
 
-public class InterfaceTileEntity extends UniversalInterfaceAdapter {
+public class InterfaceTileEntity extends UniversalInterfaceAdapter<TileEntity> {
 
     private TileEntity blockEntity = null;
     private ItemLinker.WorldPos tilePos = null;
@@ -29,18 +31,23 @@ public class InterfaceTileEntity extends UniversalInterfaceAdapter {
     @Override
     @SideOnly(Side.CLIENT)
     public void renderInventory(int tickCount, double par1, double par3, double par5, float par7) {
+        if (blockEntity != null) {
+            IBlockState state = blockEntity.getWorld().getBlockState(blockEntity.getPos());
+
+            if (state.getRenderType() == EnumBlockRenderType.MODEL) {
+                UniversalInterfaceAdapter.renderBlock(state, tickCount);
+                return;
+            }
+        }
+
         UniversalInterfaceAdapter.defaultRenderInventory(tickCount);
     }
 
     @Override
-    public boolean setLink(Object link, EntityPlayer linker) {
-        if (link instanceof TileEntity) {
-            this.blockEntity = (TileEntity) link;
-            tilePos = new ItemLinker.WorldPos(this.blockEntity.getWorld(), this.blockEntity.getPos());
-            return true;
-        }
-
-        return false;
+    public boolean setLink(TileEntity link, EntityPlayer linker) {
+        this.blockEntity = link;
+        tilePos = new ItemLinker.WorldPos(this.blockEntity.getWorld(), this.blockEntity.getPos());
+        return true;
     }
 
     @Override
@@ -51,6 +58,8 @@ public class InterfaceTileEntity extends UniversalInterfaceAdapter {
     @Override
     public void readFromNBT(NBTTagCompound nbt) {
         this.tilePos = ItemLinker.WorldPos.readFromNbt(nbt.getCompoundTag("pos"));
+
+        System.out.println(tilePos);
     }
 
     @Override
@@ -74,9 +83,6 @@ public class InterfaceTileEntity extends UniversalInterfaceAdapter {
 
     @Override
     public void onTick(int tick) {
-        if (isRemote()) {
-            return;
-        }
 
         if (blockEntity == null) {
             if (tilePos == null) {
@@ -87,8 +93,10 @@ public class InterfaceTileEntity extends UniversalInterfaceAdapter {
             this.blockEntity = tilePos.world.getTileEntity(tilePos.pos);
         }
 
-        if (blockEntity == null || this.blockEntity.isInvalid()) {
-            this.getInterface().unlink();
+        if (!isRemote()) {
+            if (blockEntity == null || this.blockEntity.isInvalid()) {
+                this.getInterface().unlink();
+            }
         }
     }
